BACKEND/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
boto3
python-dotenv
pydantic